+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.0] - 2026-04-20
 ### Added
 - Option to select Toast Position
 - Serialized Dictionary package dependency
@@ -37,7 +38,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/ToastSystem/compare/1.2.0...main
+[Unreleased]: https://github.com/1mbitshorde/ToastSystem/compare/1.3.0...main
+[1.3.0]: https://github.com/1mbitshorde/ToastSystem/tree/1.3.0/
 [1.2.0]: https://github.com/1mbitshorde/ToastSystem/tree/1.2.0/
 [1.1.0]: https://github.com/1mbitshorde/ToastSystem/tree/1.1.0/
 [1.0.0]: https://github.com/1mbitshorde/ToastSystem/tree/1.0.0/